requirements.txt
numpy
scipy
jupyterlab
ipython
matplotlib
halo
Pillow
imageio
sqlalchemy>=2
psycopg2-binary
python-dotenv
pandas
xlrd
openpyxl
Flask>=2
GraphQL-Server
graphene>=3.0
graphene_sqlalchemy
